Dev spring0

+++ backend/src/main/java/ar/edu/unq/turnero/spring/controller/DTOs/MiniTurnoDTO.kt
@@ -7,8 +7,8 @@ import ar.edu.unq.turnero.modelo.exception.EspecialidadVacioException
 class MiniTurnoDTO(
     var id: Long?,
     var nombreYApellidoPaciente:String?,
-    var dniPaciente: Int?,
-    var telefonoPaciente: Int?,
+    var dniPaciente: Long?,
+    var telefonoPaciente: Long?,
     var emailPaciente: String?,
     var fechaYHora: String?,
     var fechaEmitido: String?,
